requirements.txt
panda
numpy
tensorflow
yfinance
tweepy
google.cloud
dotenv
pytz
